Condense projectile module into single file
