+++ src/combat/mod.rs
@@ -8,5 +8,7 @@ pub mod projectile;
 pub mod shield;
 pub mod status_effects;
 
+// These exist just to reduce stutter
 pub use health::Health;
 pub use mana::Mana;
+pub use projectile::Projectile;
